gke-api/requirements.txt
flask==3.1.1
anthropic==0.43.0
google-cloud-secret-manager==2.22.0
requests==2.32.3
